Fix cache handling and return back from prefs

+++ src-tauri/types/src/cache.rs
@@ -14,5 +14,5 @@ pub struct CacheModel {
     pub id: Option<i32>,
     pub url: String,
     pub blob: Vec<u8>,
-    pub expires: i32,
+    pub expires: i64,
 }
