+++ src-tauri/types/src/cache_schema.rs
@@ -5,6 +5,6 @@ diesel::table! {
         id -> Nullable<Integer>,
         url -> Text,
         blob -> Binary,
-        expires -> Integer,
+        expires -> BigInt,
     }
 }
